Possible undefined data fixed
Some crashes shown up due to undefined data in TCPnode status request, I fixed it adding a validation for undefined (will check later on if this worked)

+++ api.js
@@ -185,11 +185,13 @@ const getTCPNodeStatus = (host, port = 8080) => {
     });
 
     client.on("end", function(){
-        lastNodeStatus = data.toString();
-        nodeResults.push(stringToNodeInfo(data.toString(), host, port));
         pendingNodes--;
-        if(pendingNodes == 0) Consensus();
-        client.destroy();
+        if(data != undefined){
+            lastNodeStatus = data.toString();
+            nodeResults.push(stringToNodeInfo(data.toString(), host, port));
+            if(pendingNodes == 0) Consensus();
+            client.destroy();
+        }
     });
 }
